project_3/chat-service-lambda/requirements.txt
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
pydantic==2.11.5
python-jose==3.4.0
requests==2.32.3
boto3==1.38.27
botocore==1.38.27